+++ src/providers/VoiceFocusProvider/index.tsx
@@ -67,8 +67,11 @@ const VoiceFocusProvider: React.FC<Props> = ({
   });
 
   const addVoiceFocus = async (device: Device): Promise<Device | VoiceFocusTransformDevice> => {
+    console.info('Add Amazon Voice Focus to the following audio input device', device);
+
     if (voiceFocusDevice) {
       const vf = await voiceFocusDevice.chooseNewInnerDevice(device);
+      console.info('Re-used the same internal state to create an Amazon Voice Focus transform device.');
       setVoiceFocusDevice(vf);
       return vf;
     }
@@ -82,6 +85,7 @@ const VoiceFocusProvider: React.FC<Props> = ({
       const transformer = await getVoiceFocusDeviceTransformer();
       const vf = await transformer?.createTransformDevice(device);
       if (vf) {
+        console.info('Created a new Amazon Voice Focus transform device.');
         setVoiceFocusDevice(vf);
         return vf;
       }
@@ -167,7 +171,7 @@ const VoiceFocusProvider: React.FC<Props> = ({
 
   useEffect(() => {
     if (voiceFocusDevice) {
-      console.info('Amazon Voice Focus transform device: ', voiceFocusDevice);
+      console.info('Current Amazon Voice Focus transform device: ', voiceFocusDevice);
     } 
   }, [voiceFocusDevice]);